io:proteoscape: Rename; use correct score
